requirements.txt
scikit-learn==0.24.2
pandas==1.2.4
streamlit==1.8.0
matplotlib==3.4.2
seaborn==0.11.1
numpy==1.20.3
deepchecks==0.5.0
joblib~=1.1.0
Pillow~=9.0.1